+++ gui/extension/CHANGELOG.md
@@ -20,7 +20,7 @@
 - BUG#38260801 Unexpected error when generating the SDK for the default language
 - BUG#38257852 MySQL REST Service generates error code with a function without parameters
 - BUG#38251258 Impossible to generate the SDK using the mrs object in MySQL Shell
--
+- BUG#38296701 Connection editor ignores additional options in the Advanced Tab
 
 ## Changes in 1.19.15+9.4.0